+++ src/NEventStore.Persistence.MsSql.Tests/PersistenceEngineFixture.cs
@@ -1,20 +1,20 @@
 ﻿using NEventStore.Persistence.Sql.Tests;
 
-namespace NEventStore.Persistence.AcceptanceTests
-{
+namespace NEventStore.Persistence.AcceptanceTests {
     using NEventStore.Persistence.Sql;
     using NEventStore.Persistence.Sql.SqlDialects;
     using NEventStore.Serialization;
     using System.Data.SqlClient;
     using System.Transactions;
 
-    public partial class PersistenceEngineFixture
-    {
+    public partial class PersistenceEngineFixture {
+        public ISqlDialect SqlDialect { get; set; } = new MsSqlDialect();
+
         /// <summary>
         /// this mimic the current NEventStore default values which is run outside any transaction (creates a scope that
         /// suppresses any transaction)
         /// </summary>
-        public TransactionScopeOption? ScopeOption { get; set; } = null; // the old default: TransactionScopeOption.Suppress;
+        public TransactionScopeOption? ScopeOption { get; set; }  // the old default: TransactionScopeOption.Suppress;
 
         public PersistenceEngineFixture()
         {
@@ -22,7 +22,7 @@ namespace NEventStore.Persistence.AcceptanceTests
             _createPersistence = pageSize =>
                 new SqlPersistenceFactory(new EnviromentConnectionFactory("MsSql", "System.Data.SqlClient"),
                     new BinarySerializer(),
-                    new MsSqlDialect(),
+                    SqlDialect,
                     pageSize: pageSize,
                     scopeOption: ScopeOption
                     ).Build();
@@ -30,7 +30,7 @@ namespace NEventStore.Persistence.AcceptanceTests
             _createPersistence = pageSize =>
                 new SqlPersistenceFactory(new EnviromentConnectionFactory("MsSql", SqlClientFactory.Instance),
                     new BinarySerializer(),
-                    new MsSqlDialect(),
+                    SqlDialect,
                     pageSize: pageSize,
                     scopeOption: ScopeOption
                     ).Build();